Extract and clean window popup code
